Hope
Im Conor Dewayne Bridges im looking for something i like doing and can make good money on for the rest of my life. Im looking for some kind of technology work and learning as much as i can. I just wanna try to make something out of my life and be able to support anyone in a relationship.